ad9517: rename files using lowercase to comply to linux standards
No functional change.
